Use no-conflict mode
When trying to use this with a Wordpress site, I learned that Wordpress
loads in "no-conflict" mode to prevent compatibility problems that could
arise from other javascript libraries.  This means the $ shortcut is not
available.

Not sure if this is the best way to do things for a jquery library like
this, but it works for now.